Refactored inner classes into their own objects
