+++ zeppelin-zengine/src/main/java/com/teragrep/zep_01/notebook/NoteManager.java
@@ -18,7 +18,6 @@
 
 package com.teragrep.zep_01.notebook;
 
-import com.teragrep.stb_01.Stubable;
 import com.teragrep.zep_01.notebook.exception.NoteNotFoundException;
 import org.apache.commons.lang3.StringUtils;
 import com.teragrep.zep_01.conf.ZeppelinConfiguration;
@@ -30,8 +29,6 @@ import org.slf4j.LoggerFactory;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -385,345 +382,6 @@ public class NoteManager {
   private String getNoteName(String notePath) {
     int pos = notePath.lastIndexOf('/');
     return notePath.substring(pos + 1);
-  }
-
-  /**
-   * Represent one folder that could contains sub folders and note files.
-   */
-  public static class Folder {
-
-    private String name;
-    private Folder parent;
-    private NotebookRepo notebookRepo;
-
-    // noteId -> NoteNode
-    private Map<String, NoteNode> notes = new HashMap<>();
-    // folderName -> Folder
-    private Map<String, Folder> subFolders = new HashMap<>();
-
-    public Folder(String name, NotebookRepo notebookRepo) {
-      this.name = name;
-      this.notebookRepo = notebookRepo;
-    }
-
-    public Folder(String name, Folder parent, NotebookRepo notebookRepo) {
-      this(name, notebookRepo);
-      this.parent = parent;
-    }
-
-    public synchronized Folder getOrCreateFolder(String folderName) {
-      if (StringUtils.isBlank(folderName)) {
-        return this;
-      }
-      if (!subFolders.containsKey(folderName)) {
-        subFolders.put(folderName, new Folder(folderName, this, notebookRepo));
-      }
-      return subFolders.get(folderName);
-    }
-
-    public Folder getParent() {
-      return parent;
-    }
-
-    public void setParent(Folder parent) {
-      this.parent = parent;
-    }
-
-    public String getName() {
-      return name;
-    }
-
-    public void setName(String name) {
-      this.name = name;
-    }
-
-    public Folder getFolder(String folderName) {
-      return subFolders.get(folderName);
-    }
-
-    public Map<String, Folder> getFolders() {
-      return subFolders;
-    }
-
-    public NoteNode getNote(String noteId) {
-      // Search for the note within this Folder
-      NoteNode rv = new NoteNodeStub();
-      if(this.notes.containsKey(noteId)){
-        rv = notes.get(noteId);
-      }
-      else {
-        for (Folder subfolder : subFolders.values()) {
-            // If a match is found in one of the subfolders, return it.
-            NoteNode noteNode = subfolder.getNote(noteId);
-            if(!noteNode.isStub()){
-              rv = noteNode;
-              break;
-            }
-          }
-        }
-      return rv;
-    }
-
-    public void addNote(String noteId, Note note) {
-      notes.put(noteId, new NoteNodeImpl(note, this, notebookRepo));
-    }
-
-    /**
-     * Attach another folder under this folder, this is used when moving folder.
-     * The path of notes under this folder also need to be updated.
-     */
-    public void addFolder(String folderName, Folder folder) throws IOException {
-      subFolders.put(folderName, folder);
-      folder.setParent(this);
-      folder.setName(folderName);
-      for (NoteNode noteNode : folder.getNoteNodeRecursively()) {
-        noteNode.updateNotePath();
-      }
-    }
-
-
-    /**
-     * Attach note under this folder, this is used when moving note
-     * @param noteNode
-     */
-    public void addNoteNode(NoteNode noteNode) {
-      this.notes.put(noteNode.getNoteId(), noteNode);
-      noteNode.setParent(this);
-    }
-
-    public void removeNote(String noteId) {
-      this.notes.remove(noteId);
-    }
-
-    public List<Note> removeFolder(String folderName,
-                                   AuthenticationInfo subject) throws IOException {
-      Folder folder = this.subFolders.remove(folderName);
-      return folder.getRawNotesRecursively();
-    }
-
-    public List<Note> getRawNotesRecursively() {
-      List<Note> notesInfo = new ArrayList<>();
-      for (NoteNode noteNode : this.notes.values()) {
-        notesInfo.add(noteNode.getRawNote());
-      }
-      for (Folder folder : subFolders.values()) {
-        notesInfo.addAll(folder.getRawNotesRecursively());
-      }
-      return notesInfo;
-    }
-
-    public List<NoteNode> getNoteNodeRecursively() {
-      List<NoteNode> noteNodeRecursively = new ArrayList<>();
-      noteNodeRecursively.addAll(this.notes.values());
-      for (Folder folder : subFolders.values()) {
-        noteNodeRecursively.addAll(folder.getNoteNodeRecursively());
-      }
-      return noteNodeRecursively;
-    }
-
-    public Map<String, NoteNode> getNotes() {
-      return notes;
-    }
-
-    public String getPath() {
-      // root
-      if (name.equals("/")) {
-        return name;
-      }
-      // folder under root
-      if (parent.name.equals("/")) {
-        return "/" + name;
-      }
-      // other cases
-      return parent.toString() + "/" + name;
-    }
-
-    @Override
-    public String toString() {
-      return getPath();
-    }
-  }
-
-  public interface NoteNode extends Stubable {
-
-    public Note getNote() throws IOException;
-    public Note getNote(boolean reload) throws IOException;
-    public String getNoteId();
-    public String getNoteName();
-    public String getNotePath();
-    public Note getRawNote();
-    public Folder getParent();
-    public void setParent(Folder parent);
-    public void setNotePath(String notePath);
-    public void updateNotePath();
-  }
-  
-  /**
-   * One node in the file system tree structure which represent the note.
-   * This class has 2 usage scenarios:
-   * 1. metadata of note (only noteId and note name is loaded via reading the file name)
-   * 2. the note object (note content is loaded from NotebookRepo)
-   *
-   * It will load note from NotebookRepo lazily until method getNote is called.
-   */
-  public static class NoteNodeImpl implements NoteNode {
-
-    private Folder parent;
-    private Note note;
-    private NotebookRepo notebookRepo;
-    public NoteNodeImpl(Note note, Folder parent, NotebookRepo notebookRepo) {
-      this.note = note;
-      this.parent = parent;
-      this.notebookRepo = notebookRepo;
-    }
-
-    @Override
-    public synchronized Note getNote() throws IOException {
-        return getNote(false);
-    }
-
-    /**
-     * This method will load note from NotebookRepo. If you just want to get noteId, noteName or
-     * notePath, you can call method getNoteId, getNoteName & getNotePath
-     * @return
-     * @throws IOException
-     */
-
-    @Override
-    public synchronized Note getNote(boolean reload) throws IOException {
-      if (!note.isLoaded() || reload) {
-        note = notebookRepo.get(note.getId(), note.getPath(), AuthenticationInfo.ANONYMOUS);
-        if (parent.toString().equals("/")) {
-          note.setPath("/" + note.getName());
-        } else {
-          note.setPath(parent.toString() + "/" + note.getName());
-        }
-        note.setCronSupported(ZeppelinConfiguration.create());
-        note.setLoaded(true);
-      }
-      return note;
-    }
-
-    @Override
-    public String getNoteId() {
-      return this.note.getId();
-    }
-
-    @Override
-    public String getNoteName() {
-      return this.note.getName();
-    }
-
-    @Override
-    public String getNotePath() {
-      if (parent.getPath().equals("/")) {
-        return parent.getPath() + note.getName();
-      } else {
-        return parent.getPath() + "/" + note.getName();
-      }
-    }
-
-    /**
-     * This method will just return the note object without checking whether it is loaded
-     * from NotebookRepo.
-     *
-     * @return
-     */
-
-    @Override
-    public Note getRawNote() {
-      return this.note;
-    }
-
-    @Override
-    public Folder getParent() {
-      return parent;
-    }
-
-    @Override
-    public String toString() {
-      return getNotePath();
-    }
-
-    @Override
-    public void setParent(Folder parent) {
-      this.parent = parent;
-    }
-
-    @Override
-    public void setNotePath(String notePath) {
-      this.note.setPath(notePath);
-    }
-
-    /**
-     * This is called when the ancestor folder is moved.
-     */
-    @Override
-    public void updateNotePath() {
-      this.note.setPath(getNotePath());
-    }
-
-    @Override
-    public boolean isStub() {
-      return false;
-    }
-  }
-  
-  public static class NoteNodeStub implements NoteNode{
-
-    @Override
-    public Note getNote() throws IOException {
-      throw new RuntimeException("NoteNode is a stub!");
-    }
-
-    @Override
-    public Note getNote(boolean reload) throws IOException {
-      throw new RuntimeException("NoteNode is a stub!");
-    }
-
-    @Override
-    public String getNoteId() {
-      throw new RuntimeException("NoteNode is a stub!");
-    }
-
-    @Override
-    public String getNoteName() {
-      throw new RuntimeException("NoteNode is a stub!");
-    }
-
-    @Override
-    public String getNotePath() {
-      throw new RuntimeException("NoteNode is a stub!");
-    }
-
-    @Override
-    public Note getRawNote() {
-      throw new RuntimeException("NoteNode is a stub!");
-    }
-
-    @Override
-    public Folder getParent() {
-      throw new RuntimeException("NoteNode is a stub!");
-    }
-
-    @Override
-    public void setParent(Folder parent) {
-      throw new RuntimeException("NoteNode is a stub!");
-    }
-
-    @Override
-    public void setNotePath(String notePath) {
-      throw new RuntimeException("NoteNode is a stub!");
-    }
-
-    @Override
-    public void updateNotePath() {
-      throw new RuntimeException("NoteNode is a stub!");
-    }
-    @Override
-    public boolean isStub() {
-      return true;
-    }
   }
